Cambios finales inicio, filtro

+++ database/migrations/2024_05_29_034418_create_hallazgos_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->unsignedInteger('fk_muestreo');
             $table->foreign('fk_muestreo')->references('id_muestreo')->on('muestreos');
             $table->unsignedInteger('cantidad');
-            $table->unsignedInteger('porcentaje');
+            $table->float('porcentaje',5,2);
             $table->timestamps();
             
         });
